+++ jobs/Backend/Task/Exchange.Infrastructure/Extensions/DateOnlyExtension.cs
@@ -1,0 +1,6 @@
+namespace Exchange.Infrastructure.Extensions;
+
+public static class DateOnlyExtension
+{
+    public static bool IsWeekend(this DateOnly date) => date.DayOfWeek is DayOfWeek.Saturday or DayOfWeek.Sunday;
+}
