+++ jobs/Backend/Task/Exchange.Infrastructure/Extensions/DateTimeExtension.cs
@@ -1,0 +1,6 @@
+namespace Exchange.Infrastructure.Extensions;
+
+public static class DateTimeExtension
+{
+    public static DateOnly ToDateOnly(this DateTime dateTime) => DateOnly.FromDateTime(dateTime);
+}
